requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-WTF==1.2.2
psycopg2-binary==2.9.10
gunicorn==23.0.0
python-dotenv==1.0.1
email-validator==2.2.0
itsdangerous==2.2.0
